~cleaned up testing
