+++ test/commands.test.ts
